+++ lib/stm32/h7/Makefile
@@ -46,7 +46,7 @@ OBJS += gpio_common_all.o gpio_common_f0234.o
 OBJS += pwr.o rcc.o
 OBJS += rcc_common_all.o
 OBJS += rng_common_v1.o
-OBJS += spi_common_all.o spi_common_v2.o
+OBJS += spi_common_all.o spi_common_v3.o
 OBJS += timer_common_all.o
 OBJS += usart_common_all.o usart_common_v2.o usart_common_fifos.o
 OBJS += quadspi_common_v1.o
